convert login function to typescript
